+++ src/app/app.html
@@ -1,1 +1,1 @@
-<ion-nav [root]="rootPage"></ion-nav>
+<ion-nav #myNav [root]="rootPage"></ion-nav>
