+++ src/pages/home/home.html
@@ -5,40 +5,6 @@
 </ion-header>
 
 <ion-content padding>
-  <div *ngIf="!logged; then thenBlock else elseBlock"></div>
-  <ng-template #thenBlock>
-    <div class="login-box">
-      <form (ngSubmit)="login()" #registerForm="ngForm">
-        <ion-row>
-          <ion-col>
-            <ion-list inset>
-
-              <ion-item>
-                <ion-input type="text" placeholder="Pseudo" name="username" [(ngModel)]="loginInfos.username"
-                           required></ion-input>
-              </ion-item>
-
-              <ion-item>
-                <ion-input type="password" placeholder="Mot de passe" name="password" [(ngModel)]="loginInfos.password"
-                           required></ion-input>
-              </ion-item>
-
-            </ion-list>
-          </ion-col>
-        </ion-row>
-
-        <ion-row>
-          <ion-col class="signup-col">
-            <button ion-button class="submit-btn" full type="submit" [disabled]="!registerForm.form.valid">Se connecter
-            </button>
-          </ion-col>
-        </ion-row>
-
-      </form>
-      <button ion-button class="register-btn" block clear (click)="createAccount()">Créer un compte</button>
-    </div>
-  </ng-template>
-  <ng-template #elseBlock>
     <div>
       <h3> Connecté en tant que {{user.username}} </h3>
       <p>
@@ -52,5 +18,5 @@
     <div>
       <button ion-button color="light" (click)="logout()">Logout</button>
     </div>
-  </ng-template>
+
 </ion-content>
