+++ src/pages/login/login.scss
@@ -1,0 +1,3 @@
+page-login {
+
+}
