+++ src/pages/registration/registration.html
@@ -1,7 +1,7 @@
 <ion-header>
 
   <ion-navbar>
-    <ion-title>registration</ion-title>
+    <ion-title>Inscription</ion-title>
   </ion-navbar>
 
 </ion-header>
